requirements.txt
torch~=2.7.0+cu126
pandas~=2.2.3
numpy~=2.2.6
scipy~=1.15.3
scikit-learn~=1.6.1
tqdm~=4.67.1
pyyaml~=6.0.2